First implementation of persistent login
Initial implementation of persistent login. Passworddetect script checks
if user data already exists in Chrome storage. If it does, the extension
builds the trie and automatically enables. If no data exists, the
extension switches to insecure mode and will prompt the user to log in
when they click the icon.